requirements.txt
streamlit>=1.28.0
fpdf2>=2.7.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
pdf2image>=1.16.0
ezdxf>=1.0.0
pdfrw>=0.4
streamlit-drawable-canvas>=0.9.0
